+++ packages/renderer/src/lib/context/ContextUtils.ts
@@ -1,0 +1,31 @@
+/**********************************************************************
+ * Copyright (C) 2023 Red Hat, Inc.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ *
+ * SPDX-License-Identifier: Apache-2.0
+ ***********************************************************************/
+import { ContextUI } from '/@/lib/context/context';
+
+function getObjectProperty<T, K extends keyof T>(obj: T, key: K): T[K] {
+  return obj[key];
+}
+
+export function transformObjectToContext<T>(obj: T, prefix: string | undefined = undefined): ContextUI {
+  // Transform our generic object as context
+  const context = new ContextUI();
+  for (const key in obj) {
+    context.setValue(prefix ? `${prefix}:${key}` : key, getObjectProperty(obj, key as keyof T));
+  }
+  return context;
+}
